[ADD] Created action and reducer for language list
